Move MetricSet to metricset.py
This is consistent with the other container classes being in separate
modules from the item class.
